pyproject.toml
[tool.poetry]
name = "software-defect-prediction"
version = "0.1.0"
description = "MLOps project for software defects prediction"
authors = ["Ekaterina Dlinnova <email@example.com>"]

packages = [{include = "software-defect-prediction"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
pytorch-lightning = "^2.0.0"
hydra-core = "^1.3.0"
kaggle = "1.6.3"
mlflow = "^2.4.0"
onnx= "^1.15.0"
onnxruntime= "^1.16.3"
onnxscript = "^0.4.0"
tensorboard="^2.0.0"
#tensorrt="^10.6.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
seaborn = "^0.13.2"
tritonclient = "^2.45.0"
dvc = "^3.0.0"
pre-commit = "^3.3.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
optuna = "^3.3.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.0"
sphinx-rtd-theme = "^1.3.0"
myst-parser = "^2.0.0"
sphinx-autodoc-typehints = "^1.25.0"
sphinx-copybutton = "^0.5.0"
sphinx-design = "^0.5.0"
sphinxcontrib-mermaid = "^0.9.0"
autodocsumm = "^0.2.11"
sphinx-autoapi = "^2.1.0"

[tool.poetry.group.dev.dependencies]
docker = "^6.1"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^24.10.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90

[tool.isort]
profile = "black"
